plugins/scry/pyproject.toml
[project]
name = "scry"
version = "1.2.9"
description = "Media generation capabilities for terminal recordings (VHS), browser recordings (Playwright), GIF processing, and media composition"
authors = [{name = "Alex Thola", email = "alexthola@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "mypy>=1.11.0",
    "pyyaml>=6.0.0",
    "types-pyyaml>=6.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require ffmpeg/vhs)",
    "slow: Slow running tests",
    "bdd: BDD-style tests with Given/When/Then",
    "requires_vhs: mark test as requiring vhs to be installed",
    "requires_ffmpeg: mark test as requiring ffmpeg to be installed",
    "requires_playwright: mark test as requiring playwright to be installed",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Relax strictness for test files
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
warn_return_any = false
